+++ crates/expo-parser/src/types.rs
@@ -38,12 +38,15 @@ impl Parser {
         self.expect(&TokenKind::LParen);
 
         let mut params = Vec::new();
+        let mut param_is_move = Vec::new();
         if !self.at(&TokenKind::RParen) {
+            param_is_move.push(self.eat(&TokenKind::Move).is_some());
             params.push(self.parse_type_expr());
             while self.eat(&TokenKind::Comma).is_some() {
                 if self.at(&TokenKind::RParen) {
                     break;
                 }
+                param_is_move.push(self.eat(&TokenKind::Move).is_some());
                 params.push(self.parse_type_expr());
             }
         }
@@ -53,6 +56,7 @@ impl Parser {
 
         TypeExpr::Function {
             params,
+            param_is_move,
             return_type: Box::new(return_type),
             span: self.span_from(start),
         }